Move matrics catalog out of submenu
