+++ packages/frontend/src/components/NavBar/BrowseMenu.tsx
@@ -6,13 +6,9 @@ import {
     IconFolders,
     IconLayoutDashboard,
 } from '@tabler/icons-react';
-import { useCallback, type FC } from 'react';
+import { type FC } from 'react';
 import { Link } from 'react-router';
-import { useProject } from '../../hooks/useProject';
 import { useSpaceSummaries } from '../../hooks/useSpaces';
-import useTracking from '../../providers/Tracking/useTracking';
-import { Hash } from '../../svgs/metricsCatalog';
-import { EventName } from '../../types/Events';
 import MantineIcon from '../common/MantineIcon';
 
 interface Props {
@@ -24,21 +20,6 @@ const BrowseMenu: FC<Props> = ({ projectUuid }) => {
         projectUuid,
         true,
     );
-
-    const { data: project } = useProject(projectUuid);
-    const { track } = useTracking();
-
-    const handleMetricsCatalogClick = useCallback(() => {
-        if (project) {
-            track({
-                name: EventName.METRICS_CATALOG_CLICKED,
-                properties: {
-                    organizationId: project.organizationUuid,
-                    projectId: projectUuid,
-                },
-            });
-        }
-    }, [project, projectUuid, track]);
 
     return (
         <Menu
@@ -83,15 +64,6 @@ const BrowseMenu: FC<Props> = ({ projectUuid }) => {
                     icon={<MantineIcon icon={IconChartAreaLine} />}
                 >
                     All saved charts
-                </Menu.Item>
-
-                <Menu.Item
-                    component={Link}
-                    to={`/projects/${projectUuid}/metrics`}
-                    icon={<Hash />}
-                    onClick={handleMetricsCatalogClick}
-                >
-                    Metrics Catalog
                 </Menu.Item>
 
                 {isInitialLoading || (spaces && spaces.length > 0) ? (
